Convert index.html to dynamic rendering
-Update CSS
-Use <Head> component
-Update routes
