--- public/index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Panda</title>
-    <link rel="stylesheet" href="css/index.css" />
-    <script src="/js/htmx.min.js"></script>
-  </head>
-
-  <body>
-    <img src="/images/sleeping_bear_logo.svg" alt="Sleeping Bear Logo" />
-    <nav>
-      <a href="/login">Login</a>
-      <a href="/about">About</a>
-      <a href="/create-club">Create Club</a>
-      <button
-        hx-post="/api/users/logout"
-        hx-include="none"
-        hx-trigger="click"
-        hx-on::after-request="window.location.href='/login'"
-      >
-        Logout
-      </button>
-    </nav>
-    <h1>Welcome to Panda</h1>
-    <p>Your app is running!</p>
-  </body>
-</html>
